little change

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: aroualid <aroualid@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 15:56:56 by aroualid          #+#    #+#             */
-/*   Updated: 2024/02/24 14:19:11 by aroualid         ###   ########.fr       */
+/*   Updated: 2024/02/23 17:42:23 by aroualid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_atoi(const char *str)
+int	ft_atoi(const char *str)
 {
-	int		i;
-	int		sign;
-	long	result;
+	int				i;
+	int				sign;
+	unsigned int	result;
 
 	i = 0;
 	sign = 1;
